fix: move active jobs button into actionbar
